Allow datapoint lists with several compositions
This commit fixes the generation of a DataFrame when the list of
DataPoints has compositions with multiple species. Previously, the
additional species would be ignored and there would be more column
headers than data columns. Now additional species are assigned a
value of 0.0 if they aren't present in the DataPoint.